updated the fake walmart frittte project on different table and run …

+++ FRITTTE/src/main/java/com/revature/frittte/creditcard/CreditCard.java
@@ -16,7 +16,7 @@ import javax.persistence.*;
 public class CreditCard {
     @Id
     @Column(name = "cc_number", nullable = false)
-    private int creditCardNumber;
+    private long creditCardNumber;
     @Column(name = "cc_name" ,nullable = false)
     private String creditCardName;
     private int cvv;
